made molecule abstraction runnable, shared border now gets their labels hashed
